+++ blueprints/incredible-squaring-eigenlayer/src/lib.rs
@@ -70,12 +70,11 @@ pub fn noop(_: u32) {
     event_listener(
         listener = EvmContractEventListener(
             instance = IncredibleSquaringTaskManager,
-            event = IncredibleSquaringTaskManager::NewTaskCreated,
-            event_converter = convert_event_to_inputs,
-            callback = noop,
             abi = INCREDIBLE_SQUARING_TASK_MANAGER_ABI_STRING,
         ),
         event = IncredibleSquaringTaskManager::NewTaskCreated,
+        pre_processor = convert_event_to_inputs,
+        post_processor = noop,
     ),
 )]
 pub async fn xsquare_eigen(