+++ blueprint-metadata/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.5](https://github.com/tangle-network/gadget/compare/blueprint-metadata-v0.1.4...blueprint-metadata-v0.1.5) - 2024-10-24
+
+### Other
+
+- updated the following local packages: gadget-blueprint-proc-macro-core
+
 ## [0.1.3](https://github.com/tangle-network/gadget/compare/blueprint-metadata-v0.1.2...blueprint-metadata-v0.1.3) - 2024-10-23
 
 ### Fixed