+++ macros/blueprint-proc-macro/CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.2](https://github.com/tangle-network/gadget/compare/gadget-blueprint-proc-macro-v0.2.1...gadget-blueprint-proc-macro-v0.2.2) - 2024-10-24
+
+### Other
+
+- Fix bug with tangle contexts ([#391](https://github.com/tangle-network/gadget/pull/391))
+
+## [0.2.1](https://github.com/tangle-network/gadget/compare/gadget-blueprint-proc-macro-v0.2.0...gadget-blueprint-proc-macro-v0.2.1) - 2024-10-24
+
+### Other
+
+- Event Flows for Tangle ([#363](https://github.com/tangle-network/gadget/pull/363))
+
 ## [0.2.0](https://github.com/tangle-network/gadget/compare/gadget-blueprint-proc-macro-v0.1.2...gadget-blueprint-proc-macro-v0.2.0) - 2024-10-23
 
 ### Added